# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: NadiKota API
  description: |
    API untuk platform pelaporan infrastruktur Kota Batam.
    Semua endpoint diawali dengan /api/v1.
  version: 1.0.0
  contact:
    name: NadiKota Team
servers:
  - url: http://localhost:8000/api/v1
    description: Local development
security:
  - bearerAuth: []
paths:
...
  /tickets:
    get:
      summary: Antrean tiket (admin)
      tags: [Tickets]
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Daftar tiket
  /tickets/{ticket}:
    get:
      summary: Detail tiket
      tags: [Tickets]
      parameters:
        - name: ticket
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Detail tiket
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Sanctum
...